+++ DotSdkSamples/Samples/FaceAutoCapture/FaceAutoCaptureSampleResult.swift
@@ -6,14 +6,12 @@ class FaceAutoCaptureSampleResult: SampleResult {
     let image: UIImage
     let resultDescription: String? = nil
     let confidence: Double
-    let passiveLivenessFaceAttribute: FaceAttribute
     let faceAspects: FaceAspects
     let faceQuality: FaceQuality
 
-    init(image: UIImage, confidence: Double, passiveLivenessFaceAttribute: FaceAttribute, faceAspects: FaceAspects, faceQuality: FaceQuality) {
+    init(image: UIImage, confidence: Double, faceAspects: FaceAspects, faceQuality: FaceQuality) {
         self.image = image
         self.confidence = confidence
-        self.passiveLivenessFaceAttribute = passiveLivenessFaceAttribute
         self.faceAspects = faceAspects
         self.faceQuality = faceQuality
     }
@@ -23,7 +21,6 @@ extension FaceAutoCaptureSampleResult: Encodable {
     
     enum Keys: String, CodingKey {
         case confidence
-        case passiveLivenessFaceAttribute
         case faceAspects
         case faceQuality
     }
@@ -31,7 +28,6 @@ extension FaceAutoCaptureSampleResult: Encodable {
     public func encode(to encoder: Encoder) throws {
         var container = encoder.container(keyedBy: Keys.self)
         try container.encode(confidence, forKey: .confidence)
-        try container.encode(passiveLivenessFaceAttribute, forKey: .passiveLivenessFaceAttribute)
         try container.encode(faceAspects, forKey: .faceAspects)
         try container.encode(faceQuality, forKey: .faceQuality)
     }