+++ DotSdkSamples/SamplesViewController.swift
@@ -16,14 +16,12 @@ class SamplesViewController: UIViewController {
     private lazy var faceAutoCaptureButton = buttonFactory.create(.faceAutoCapture)
     private lazy var smileLivenessButton = buttonFactory.create(.smileLiveness)
     private lazy var magnifEyeLivenessButton = buttonFactory.create(.magnifEyeLiveness)
-    private lazy var faceMatcherButton = buttonFactory.create(.faceMatcher)
     
     private lazy var buttons: [UIButton] = [documentAutoCaptureButton,
                                             nfcReadingButton,
                                             faceAutoCaptureButton,
                                             smileLivenessButton,
-                                            magnifEyeLivenessButton,
-                                            faceMatcherButton]
+                                            magnifEyeLivenessButton]
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -76,13 +74,6 @@ extension SamplesViewController {
         let viewController = SmileLivenessContainerViewController()
         navigationController?.pushViewController(viewController, animated: true)
     }
-    
-    @objc private func showFaceMatcher() {
-        guard checkDotSdkIsInitialized() else { return }
-        
-        let viewController = FaceMatchingViewController()
-        navigationController?.pushViewController(viewController, animated: true)
-    }
 }
 
 extension SamplesViewController {
@@ -124,7 +115,6 @@ extension SamplesViewController {
         faceAutoCaptureButton.addTarget(self, action: #selector(showFaceAutoCapture), for: .touchUpInside)
         smileLivenessButton.addTarget(self, action: #selector(showSmileLiveness), for: .touchUpInside)
         magnifEyeLivenessButton.addTarget(self, action: #selector(showMagnifEyeLiveness), for: .touchUpInside)
-        faceMatcherButton.addTarget(self, action: #selector(showFaceMatcher), for: .touchUpInside)
         
         NSLayoutConstraint.activate([
             scrollView.frameLayoutGuide.topAnchor.constraint(equalTo: view.topAnchor),
@@ -147,8 +137,7 @@ extension SamplesViewController {
             nfcReadingButton.heightAnchor.constraint(equalToConstant: buttonFactory.getHeight(.nfcReading)),
             faceAutoCaptureButton.heightAnchor.constraint(equalToConstant: buttonFactory.getHeight(.faceAutoCapture)),
             smileLivenessButton.heightAnchor.constraint(equalToConstant: buttonFactory.getHeight(.smileLiveness)),
-            magnifEyeLivenessButton.heightAnchor.constraint(equalToConstant: buttonFactory.getHeight(.magnifEyeLiveness)),
-            faceMatcherButton.heightAnchor.constraint(equalToConstant: buttonFactory.getHeight(.faceMatcher)),
+            magnifEyeLivenessButton.heightAnchor.constraint(equalToConstant: buttonFactory.getHeight(.magnifEyeLiveness))
         ])
     }
 }